Move subnavbarlink component to same folder
